moved attachments pdf serializer & comma separated serializer field to their places
